+++ lisa-examples/src/main/scala/Exercise.scala
@@ -1,7 +1,6 @@
 import lisa.automation.kernel.OLPropositionalSolver.*
 import lisa.automation.kernel.SimpleSimplifier.*
 import lisa.kernel.proof.SequentCalculus.*
-import lisa.utils.Library
 import lisa.utils.Printer
 
 object Exercise extends lisa.Main {
