+++ lisa-utils/src/main/scala/lisa/prooflib/Exports.scala
@@ -1,0 +1,6 @@
+package lisa.prooflib
+
+object Exports {
+  export BasicStepTactic.*
+  export lisa.prooflib.SimpleDeducedSteps.*
+}
