+++ lisa-utils/src/main/scala/lisa/prooflib/SimpleDeducedSteps.scala
@@ -1,4 +1,4 @@
-package lisa.utils.tactics
+package lisa.prooflib
 
 import lisa.kernel.fol.FOL
 import lisa.kernel.proof.SCProof
@@ -7,13 +7,12 @@ import lisa.kernel.proof.SequentCalculus.RewriteTrue
 import lisa.kernel.proof.SequentCalculus.SCProofStep
 import lisa.kernel.proof.SequentCalculus.Sequent
 import lisa.kernel.proof.SequentCalculus as SC
+import lisa.prooflib.ProofTacticLib.{_, given}
+import lisa.prooflib.*
 import lisa.utils.FOLParser
-import lisa.utils.Helpers.{_, given}
-import lisa.utils.Library
-import lisa.utils.LisaException
-import lisa.utils.OutputManager
+import lisa.utils.KernelHelpers.{_, given}
 import lisa.utils.Printer
-import lisa.utils.tactics.ProofTacticLib.{_, given}
+
 object SimpleDeducedSteps {
 
   object Restate extends ProofTactic with ParameterlessHave with ParameterlessAndThen {
