+++ lisa-utils/src/test/scala/lisa/kernel/EquivalenceCheckerTests.scala
@@ -3,9 +3,9 @@ package lisa.kernel
 import lisa.kernel.fol.FOL
 import lisa.kernel.fol.FOL.*
 import lisa.utils.FOLPrinter
-import lisa.utils.Helpers._
-import lisa.utils.Helpers.given_Conversion_Identifier_String
-import lisa.utils.Helpers.given_Conversion_String_Identifier
+import lisa.utils.KernelHelpers._
+import lisa.utils.KernelHelpers.given_Conversion_Identifier_String
+import lisa.utils.KernelHelpers.given_Conversion_String_Identifier
 import org.scalatest.funsuite.AnyFunSuite
 
 import scala.collection.MapView
