+++ lisa-utils/src/test/scala/lisa/kernel/InvalidProofPathTests.scala
@@ -4,7 +4,7 @@ import lisa.kernel.proof.SCProofCheckerJudgement.SCInvalidProof
 import lisa.kernel.proof.SequentCalculus.*
 import lisa.kernel.proof.*
 import lisa.test.ProofCheckerSuite
-import lisa.utils.Helpers.{_, given}
+import lisa.utils.KernelHelpers.{_, given}
 
 class InvalidProofPathTests extends ProofCheckerSuite {
   def checkPath(invalidProof: SCProof, expectedPath: Seq[Int]): Unit = {
