+++ lisa-utils/src/test/scala/lisa/test/TestTheoryLibrary.scala
@@ -1,6 +1,6 @@
 package lisa.test
 
-import lisa.utils.Library
+import lisa.prooflib.Library
 
 object TestTheoryLibrary extends Library(TestTheory.runningTestTheory) {
   export TestTheory.*
