+++ lisa-utils/src/test/scala/lisa/utils/ProofTacticTestLib.scala
@@ -1,9 +1,10 @@
 package lisa.utils
 
-import lisa.kernel.proof.{SequentCalculus as SC}
-import lisa.utils.Library
-import lisa.utils.tactics.BasicStepTactic.*
-import lisa.utils.tactics.ProofTacticLib
+import lisa.kernel.proof.SequentCalculus as SC
+import lisa.prooflib.BasicMain
+import lisa.prooflib.BasicStepTactic.*
+import lisa.prooflib.Library
+import lisa.prooflib.ProofTacticLib
 import org.scalatest.funsuite.AnyFunSuite
 
 import scala.collection.immutable.LazyList
