+++ lisa-utils/src/test/scala/lisa/utils/TheoriesHelpersTest.scala
@@ -6,9 +6,9 @@ import lisa.kernel.proof.RunningTheoryJudgement.InvalidJustificationException
 import lisa.kernel.proof.SCProof
 import lisa.kernel.proof.SequentCalculus.Hypothesis
 import lisa.test.TestTheory
-import lisa.utils.Helpers._
-import lisa.utils.Helpers.given_Conversion_Identifier_String
-import lisa.utils.Helpers.given_Conversion_String_Identifier
+import lisa.utils.KernelHelpers._
+import lisa.utils.KernelHelpers.given_Conversion_Identifier_String
+import lisa.utils.KernelHelpers.given_Conversion_String_Identifier
 import org.scalatest.funsuite.AnyFunSuite
 
 class TheoriesHelpersTest extends AnyFunSuite {
