+++ src/main/scala/lisa/automation/kernel/OLPropositionalSolver.scala
@@ -3,19 +3,15 @@ package lisa.automation.kernel
 import lisa.kernel.fol.FOL.*
 import lisa.kernel.proof.SCProof
 import lisa.kernel.proof.SequentCalculus.*
-import lisa.kernel.proof.SequentCalculus as LK
-import lisa.utils.Helpers.{_, given}
-import lisa.utils.Helpers.{_, given}
-import lisa.utils.Library
-import lisa.utils.tactics.ProofTacticLib.ParameterlessAndThen
-import lisa.utils.tactics.ProofTacticLib.ParameterlessHave
-import lisa.utils.tactics.ProofTacticLib.ProofTactic
+import lisa.prooflib.Library
+import lisa.prooflib.ProofTacticLib.*
+import lisa.utils.KernelHelpers.{_, given}
 
 object OLPropositionalSolver {
 
   /**
    * A tactic object dedicated to solve any propositionaly provable sequent (possibly in exponential time). Can be used with arbitrary many premises.
-   * Leverages the OL algorithm for propositional logic.
+   * Leverages the OL algorithm for scalafmpropositional logic.
    */
   object Tautology extends ProofTactic with ParameterlessHave with ParameterlessAndThen {
