+++ src/main/scala/lisa/automation/kernel/ProofTactics.scala
@@ -4,7 +4,7 @@ import lisa.kernel.fol.FOL.*
 import lisa.kernel.proof.SCProof
 import lisa.kernel.proof.SequentCalculus.*
 import lisa.utils.FOLPrinter
-import lisa.utils.Helpers.{_, given}
+import lisa.utils.KernelHelpers.{_, given}
 
 /**
  * SCProof tactics are a set of strategies that help the user write proofs in a more expressive way
