+++ src/main/scala/lisa/proven/peano_example/Peano.scala
@@ -5,9 +5,9 @@ import lisa.kernel.fol.FOL.*
 import lisa.kernel.proof.RunningTheory
 import lisa.kernel.proof.SCProof
 import lisa.kernel.proof.SequentCalculus.*
-import lisa.utils.Helpers.{_, given}
-import lisa.utils.Library
-import lisa.utils.OutputManager
+import lisa.prooflib.Library
+import lisa.prooflib.OutputManager
+import lisa.utils.KernelHelpers.{_, given}
 import lisa.utils.Printer
 
 object Peano { /*
